pyproject.toml
[project]
name = "orchid"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "click>=8.3.1",
    "psutil>=7.2.2",
    "pydantic>=2.12.5",
    "pyyaml>=6.0.3",
    "textual>=8.1.1",
]

[project.scripts]
orchid = "orchid.cli:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/orchid", "src/tui"]
